[TAS-303] Replace flatMap/fold chains with either {} DSL throughout (#177)

+++ slack/domain/src/commonMain/kotlin/com/gchristov/thecodinglove/slack/domain/usecase/SlackSelfDestructUseCase.kt
@@ -1,7 +1,6 @@
 package com.gchristov.thecodinglove.slack.domain.usecase
 
 import arrow.core.Either
-import arrow.core.flatMap
 import arrow.core.raise.either
 import co.touchlab.kermit.Logger
 import com.gchristov.thecodinglove.common.kotlin.debug
@@ -23,38 +22,30 @@ internal class RealSlackSelfDestructUseCase(
     private val tag = this::class.simpleName
 
     override suspend operator fun invoke(): Either<Throwable, Unit> = withContext(dispatcher) {
-        log.debug(tag, "Fetching all self-destruct messages")
-        slackRepository
-            .getSelfDestructMessages()
-            .flatMap { messages ->
-                val messagesToDestroy = messages.filter { clock.now().toEpochMilliseconds() >= it.destroyTimestamp }
-                log.debug(tag, "Self-destructing messages: total=${messages.size}, toDestroy=${messagesToDestroy.size}")
-                messagesToDestroy
-                    .map { message ->
-                        log.debug(tag, "Obtaining user token: userId=${message.userId}")
-                        slackRepository
-                            .getAuthToken(tokenId = message.userId)
-                            .fold(
-                                ifLeft = {
-                                    log.debug(tag, "User token not found, deleting self-destruct state: message=${message.id}")
-                                    slackRepository.deleteSelfDestructMessage(message.id)
-                                },
-                                ifRight = { token ->
-                                    log.debug(tag, "Deleting Slack message: message=${message.id}")
-                                    slackRepository
-                                        .deleteMessage(
-                                            authToken = token.token,
-                                            channelId = message.channelId,
-                                            messageTs = message.messageTs,
-                                        )
-                                        .flatMap {
-                                            log.debug(tag, "Deleting self-destruct state: message=${message.id}")
-                                            slackRepository.deleteSelfDestructMessage(message.id)
-                                        }
-                                }
-                            )
+        either {
+            log.debug(tag, "Fetching all self-destruct messages")
+            val messages = slackRepository.getSelfDestructMessages().bind()
+            val messagesToDestroy = messages.filter { clock.now().toEpochMilliseconds() >= it.destroyTimestamp }
+            log.debug(tag, "Self-destructing messages: total=${messages.size}, toDestroy=${messagesToDestroy.size}")
+            messagesToDestroy.map { message ->
+                log.debug(tag, "Obtaining user token: userId=${message.userId}")
+                when (val tokenResult = slackRepository.getAuthToken(tokenId = message.userId)) {
+                    is Either.Left -> {
+                        log.debug(tag, "User token not found, deleting self-destruct state: message=${message.id}")
+                        slackRepository.deleteSelfDestructMessage(message.id)
                     }
-                    .let { l -> either { l.bindAll() } }
-            }
+                    is Either.Right -> either {
+                        log.debug(tag, "Deleting Slack message: message=${message.id}")
+                        slackRepository.deleteMessage(
+                            authToken = tokenResult.value.token,
+                            channelId = message.channelId,
+                            messageTs = message.messageTs,
+                        ).bind()
+                        log.debug(tag, "Deleting self-destruct state: message=${message.id}")
+                        slackRepository.deleteSelfDestructMessage(message.id).bind()
+                    }
+                }
+            }.bindAll()
+        }
     }
 }
